feat(models/metrics): use multilabel instead of multiclass for dice loss
This commit introduces a new torchvision.transforms class to invert the
normalisation process of dataloading in order to add train/val/test
samples to the Tensorboard instance for viewing. Additionally, this
commit also fixes a bug with the learning rate scheduler which caused
the learning rate to only increase due to an improperly set parameter.