perf(dataloading): Batch loading time decreased by 23.58%
This is due to the removed for-loops of stacking the multiple CINE
images in the __getitem__ method of the datasets, which are replaced by
built-in torch and numpy functions.

Additionally, indices loading and image concatenation methods are moved
out of the dataset classes as I do not want to implement inheritance for
something this simple for now. An additional test is included to check
whether images are rotated when loading them from the dataset or when
using cv2.