fix(models): change default to not use dilations for 5/30 frames.
There is a very odd crash that occurs on WSL2 (not sure if it happens
otherwise) when the model with dilations enabled are checkpointed for
the last epoch. Complete kernel panic and is unrecoverable, while the
in-progress checkpoint save is corrupted. Requires more investigation.
For now, this is just a band-aid fix.